requirements.txt
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
sqlparse==0.5.1
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
django-autocomplete-light==3.11.0
django-cors-headers==4.5.0
djangorestframework-simplejwt==5.3.1
idna==3.10
pillow==10.4.0
PyJWT==2.9.0
qrcode==8.0
razorpay==1.4.2
reportlab==4.2.5
requests==2.32.3
setuptools==75.1.0
urllib3==2.2.3